Refact: Move entry point logic to own file. Describe API usage
- To avoid import problems when installing using setuptools
the entrypoint is imported from package.filename.

- version updated

- README updated to describe usage using import for easier
implementation in python scripts.
