+++ package_wrapper/__init__.py
@@ -1,2 +1,2 @@
 from package_wrapper.version import __version__
-from package_wrapper.main import package_api
+
