+++ package_wrapper/version.py
@@ -1,1 +1,1 @@
-__version__ = "1.1.0"
+__version__ = "1.1.1"
